move geocoding code into its own module - there's a lot of it
